Добавлена возможность фиксировать результат работы для Владельца задачи и для исполнителя. Автоматическая подстановка текущего пользователя при заведении заявки

+++ TasksDK/ViewModel/MainViewModel.cs
@@ -33,10 +33,11 @@ namespace TasksDK.ViewModel
         private Stack<EmployeeTask> _taskStack = new Stack<EmployeeTask>();
         public Stack<EmployeeTask> TaskStack { get => _taskStack; set => _taskStack = value; }
         private ObservableCollection<Employee> _analytics = new ObservableCollection<Employee>();
-        public ObservableCollection<Employee> Analytics { get=> _analytics; set=> _analytics=value; }
+        public ObservableCollection<Employee> Analytics { get => _analytics; set => _analytics = value; }
         private ObservableCollection<string> _analytics_str = new ObservableCollection<string>();
         public ObservableCollection<string> Analytics_str { get => _analytics_str; set => _analytics_str = value; }
-
+        private Analytic _currentAnalytic;
+        private Employee _currentEmployee;
 
         /// <summary>
         /// Новая задача. Привязка из окна AddTask
@@ -46,7 +47,7 @@ namespace TasksDK.ViewModel
         private EmployeeTask _selectedTask;
         public EmployeeTask SelectedTask { get => _selectedTask; set => _selectedTask = value; }
         private ObservableCollection<Process> _selectedProcesses = new ObservableCollection<Process>();
-        public ObservableCollection<Process> SelectedProcesses { get => _selectedProcesses; set => _selectedProcesses=value; }
+        public ObservableCollection<Process> SelectedProcesses { get => _selectedProcesses; set => _selectedProcesses = value; }
 
         #endregion
 
@@ -54,9 +55,12 @@ namespace TasksDK.ViewModel
         public RelayCommand AddNewTaskCommand { get; set; }
         public RelayCommand AddNewSubTaskCommand { get; set; }
         public RelayCommand BackCommand { get; set; }
+        public RelayCommand FilterTasks_My { get; set; }
+        public RelayCommand FilterTasks_All { get; set; }
         public RelayCommand<EmployeeTask> ViewTaskCommand { get; set; }
         public RelayCommand<EmployeeTask> DeleteTask { get; set; }
         public RelayCommand<EmployeeTask> SelectParentCommand { get; set; }
+        public RelayCommand<EmployeeTask> AddResult { get; set; }
         public RelayCommand<EmployeeTask> SelectTaskCommand { get; set; }
         public RelayCommand<ICollection<object>> StoreProcessSelection { get; set; }
 
@@ -75,10 +79,16 @@ namespace TasksDK.ViewModel
         private void AddNewTask()
         {
             AddTask addTaskWindow = new AddTask();
+            newTask.Owner = _currentEmployee;
+            addTaskWindow.Reporter.IgnoreTextChange = true;
+            addTaskWindow.Reporter.Text = _currentEmployee.FIO;
+            addTaskWindow.Reporter.IgnoreTextChange = false;
             if (addTaskWindow.ShowDialog() == true)
             {
+                Console.WriteLine(newTask.Owner.FIO);
                 newTask.Reporter = Analytics.FirstOrDefault(i => i.FIO.Equals(newTask.Reporter.FIO));
                 newTask.Assignee = Analytics.FirstOrDefault(i => i.FIO.Equals(newTask.Assignee.FIO));
+                newTask.Owner = Analytics.FirstOrDefault(i => i.FIO.Equals(newTask.Reporter.FIO));
                 _tasks.AddTask(newTask.AsCopy());
                 _currentTask.ChildTasks.Add(newTask.AsCopy());
                 CurrentTasks.Add(NewTask.AsCopy());
@@ -88,9 +98,15 @@ namespace TasksDK.ViewModel
 
         private void AddSubTask()
         {
+            newTask.Owner = _currentEmployee;
+
             if (SelectedTask != null)
             {
                 AddTask addTaskWindow = new AddTask();
+                newTask.Owner = _currentEmployee;
+                addTaskWindow.Reporter.IgnoreTextChange = true;
+                addTaskWindow.Reporter.Text = _currentEmployee.FIO;
+                addTaskWindow.Reporter.IgnoreTextChange = false;
                 if (addTaskWindow.ShowDialog() == true)
                 {
                     SelectedTask.ChildTasks.Add(newTask.AsCopy());
@@ -102,14 +118,20 @@ namespace TasksDK.ViewModel
                 UpdateTasks();
 
             }
-            
+
         }
 
         private void InitializeData()
         {
             Processes = _tasks.GetProcesses();
             _mainTask.ChildTasks = new List<EmployeeTask>(_tasks.GetTasks(null));
-            foreach(Employee empl in _employees.GetEmployees())
+            _currentAnalytic = _employees.GetCurrentAnalytic();
+            _currentEmployee = new Employee
+            {
+                FIO = $"{_currentAnalytic.LastName} {_currentAnalytic.FirstName} {_currentAnalytic.FatherName}",
+                AnalyticId = _currentAnalytic.Id
+            };
+            foreach (Employee empl in _employees.GetEmployees())
             {
                 Analytics.Add(empl);
                 Analytics_str.Add(empl.FIO);
@@ -117,7 +139,7 @@ namespace TasksDK.ViewModel
             _currentTask = _mainTask;
             TaskStack.Push(_mainTask);
             CurrentTasks = new ObservableCollection<EmployeeTask>(_mainTask.ChildTasks);
-            if(_mainTask.ChildTasks.Count>0)
+            if (_mainTask.ChildTasks.Count > 0)
                 SelectedTask = _mainTask.ChildTasks[0];
         }
 
@@ -125,39 +147,107 @@ namespace TasksDK.ViewModel
         {
             AddNewTaskCommand = new RelayCommand(AddNewTask);
             AddNewSubTaskCommand = new RelayCommand(AddSubTask);
-            ViewTaskCommand = new RelayCommand<EmployeeTask>(ViewTask);
+            ViewTaskCommand = new RelayCommand<EmployeeTask>(EditTask);
             SelectParentCommand = new RelayCommand<EmployeeTask>(SelectParent);
             BackCommand = new RelayCommand(Back);
             SelectTaskCommand = new RelayCommand<EmployeeTask>(SelectTask);
             DeleteTask = new RelayCommand<EmployeeTask>(DeleteTaskMethod);
             StoreProcessSelection = new RelayCommand<ICollection<object>>(StoreProcesses);
+            FilterTasks_My = new RelayCommand(ShowMyTasks);
+            FilterTasks_All = new RelayCommand(ShowAllTAsks);
+            AddResult = new RelayCommand<EmployeeTask>(AddResultMethod);
+        }
+
+        private void AddResultMethod(EmployeeTask task)
+        {
+            ResultsWindow dialog = new ResultsWindow();
+            int participationIndex;
+            if (task.Reporter.FIO.Equals(_currentEmployee.FIO))//пользователь инициатор задачи
+            {
+                participationIndex = 1;
+            }
+            else if (task.Assignee.FIO.Equals(_currentEmployee.FIO)) //пользователь как ответственное лицо
+            {
+                participationIndex = 2;
+            }
+            else //задача не принадлежит пользователю
+            {
+                MessageBox.Show("Вы не являетесь участником данной задачи и не можете заявить результат", "Доступ запрещен", MessageBoxButton.OK, MessageBoxImage.Exclamation);
+                return;
+            }
+
+            if(participationIndex==1 && !string.IsNullOrEmpty(task.SupervisorComment))
+            {
+                dialog.CommentBox.Text = task.SupervisorComment;
+            }
+            else if (!string.IsNullOrEmpty(task.SupervisorComment))
+            {
+                dialog.CommentBox.Text = task.EmployeeComment;
+            }
+
+            if (dialog.ShowDialog() == true)
+            {
+                if (participationIndex == 1)
+                {
+                    task.SupervisorComment = dialog.CommentBox.Text;
+                    
+                }
+                else
+                {
+                    task.EmployeeComment = dialog.CommentBox.Text;
+                }
+                dialog.CommentBox.Text = string.Empty;
+                _tasks.Commit();
+            }
+        }
+
+        private void ShowMyTasks()
+        {
+            _mainTask = new EmployeeTask();
+            string fio = $"{_currentAnalytic.LastName} {_currentAnalytic.FirstName} {_currentAnalytic.FatherName}";
+            _mainTask.ChildTasks = new List<EmployeeTask>(_tasks.GetTasks().Where(i => i.Assignee.FIO != null && i.Assignee.FIO.Equals(fio) || i.Owner.FIO != null && i.Assignee.FIO.Equals(fio)));
+            _currentTask = _mainTask;
+            TaskStack.Clear();
+            TaskStack.Push(_mainTask);
+            if (_mainTask.ChildTasks.Count > 0)
+                SelectedTask = _mainTask.ChildTasks[0];
+            UpdateTasks();
+        }
+
+        private void ShowAllTAsks()
+        {
+            _mainTask = new EmployeeTask();
+            _mainTask.ChildTasks = new List<EmployeeTask>(_tasks.GetTasks(null));
+            _currentTask = _mainTask;
+            TaskStack.Clear();
+            TaskStack.Push(_mainTask);
+            if (_mainTask.ChildTasks.Count > 0)
+                SelectedTask = _mainTask.ChildTasks[0];
+            UpdateTasks();
         }
 
         private void StoreProcesses(ICollection<object> obj)
         {
             List<ProcessProxy> _processes = new List<ProcessProxy>();
-            foreach(object item in obj)
+            foreach (object item in obj)
             {
-                if(item is Process)
+                if (item is Process)
                 {
                     _processes.Add(new ProcessProxy
                     {
-                        ProcessId=(item as Process).Id
+                        ProcessId = (item as Process).Id
                     });
                 }
             }
             newTask.Processes = _processes;
-            
+
         }
 
         private void DeleteSubTasks(EmployeeTask task)
         {
-            if (task.ChildTasks.Count > 0)
+            while (task.ChildTasks.Count > 0)
             {
-                while(task.ChildTasks.Count>0)
-                {
-                    DeleteSubTasks(task.ChildTasks[0]);
-                }
+                DeleteSubTasks(task.ChildTasks[0]);
             }
             CurrentTasks.Remove(task);
             _tasks.Remove(task);
@@ -166,7 +256,7 @@ namespace TasksDK.ViewModel
         private void DeleteTaskMethod(EmployeeTask task)
         {
             EmployeeTask deletedTask = null;
-            foreach(EmployeeTask _task in CurrentTasks)
+            foreach (EmployeeTask _task in CurrentTasks)
             {
                 if (_task.Equals(task))
                 {
@@ -174,11 +264,11 @@ namespace TasksDK.ViewModel
                     break;
                 }
             }
-            if (deletedTask != null)
+            if (deletedTask != null && deletedTask.ChildTasks != null)
             {
-                if(deletedTask.ChildTasks!=null && deletedTask.ChildTasks.Count > 0)
+                if (deletedTask.ChildTasks.Count > 0)
                 {
-                    if(MessageBox.Show("Удаляемая задача содержит подзадачи, которые также будут удалены.\r\n Вы уверены?", "Удаление задачи", MessageBoxButton.YesNo, MessageBoxImage.Exclamation) == MessageBoxResult.Yes)
+                    if (MessageBox.Show("Удаляемая задача содержит подзадачи, которые также будут удалены.\r\n Вы уверены?", "Удаление задачи", MessageBoxButton.YesNo, MessageBoxImage.Exclamation) == MessageBoxResult.Yes)
                     {
                         DeleteSubTasks(deletedTask);
                         _currentTask.ChildTasks.Remove(deletedTask);
@@ -193,6 +283,16 @@ namespace TasksDK.ViewModel
                         return;
                     }
                 }
+                else
+                {
+                    DeleteSubTasks(deletedTask);
+                    _currentTask.ChildTasks.Remove(deletedTask);
+                    if (_currentTask.ChildTasks.Count < 1)
+                    {
+                        Back();
+                    }
+                    UpdateTasks();
+                }
 
             }
         }
@@ -202,7 +302,7 @@ namespace TasksDK.ViewModel
             SelectedTask = task;
             RaisePropertyChanged("SelectedTask");
             SelectedProcesses.Clear();
-            foreach(ProcessProxy processProxy in SelectedTask.Processes)
+            foreach (ProcessProxy processProxy in SelectedTask.Processes)
             {
                 SelectedProcesses.Add(Processes.FirstOrDefault(i => i.Id == processProxy.ProcessId));
             }
@@ -232,23 +332,27 @@ namespace TasksDK.ViewModel
         private void UpdateTasks()
         {
             CurrentTasks.Clear();
-            
+
             foreach (EmployeeTask childTask in _currentTask.ChildTasks)
             {
                 CurrentTasks.Add(childTask);
             }
         }
-        private void ViewTask(EmployeeTask task)
+
+        private void EditTask(EmployeeTask task)
         {
             EmployeeTask tempTask = task.AsCopy();
             newTask = task;
             AddTask addTaskWindow = new AddTask();
+            
+            addTaskWindow.ProcessList.SelectedItemsOverride = SelectedTask.Processes.Select(i=>Processes.FirstOrDefault(n=>n.Id==i.ProcessId)).ToList();
             if (addTaskWindow.ShowDialog() == false)
             {
                 task.CopyFields(tempTask);
             }
             else
             {
+                _tasks.Commit();
                 UpdateTasks();
             }
             newTask = new EmployeeTask();
